refactor: Modify FolderStructure
