+++ src/Main.jsx
@@ -1,0 +1,14 @@
+import { Outlet } from 'react-router-dom';
+import Navbar from './components/Navbar/Navbar';
+import Footer from './components/Footer/Footer';
+
+const Main = () => {
+  return (
+    <>
+      <Navbar />
+      <Outlet />
+      <Footer />
+    </>
+  );
+};
+export default Main;
